eproba/requirements.txt
django==5.1.6
django-crispy-forms==2.3
crispy-bulma==0.11.0
djangorestframework==3.15.2
whitenoise==6.7.0
unidecode==1.3.8
weasyprint==62.3
fcm-django==2.2.1
gunicorn==23.0.0
django-oauth-toolkit==3.0.1
pwned-passwords-django==5.1.1
pre-commit==4.1.0
django-constance==4.3.2
django-picklefield==3.2
django-maintenance-mode==0.21.1
google-api-python-client==2.154.0
python-dotenv==1.0.1
psycopg[binary]==3.2.4
drf-spectacular==0.27.2
django-dbbackup==4.2.1